feat: use cached max trace heights in leaf/internal aggregation

+++ extensions/native/circuit/src/extension.rs
@@ -475,7 +475,7 @@ impl InitFileGenerator for Rv32WithKernelsConfig {}
 // Pre-computed maximum trace heights for NativeConfig. Found by doubling
 // the actual trace heights of kitchen-sink leaf verification (except for
 // VariableRangeChecker, which has a fixed height).
-pub const NATIVE_MAX_TRACE_HEIGHTS: &[usize] = &[
+pub const NATIVE_MAX_TRACE_HEIGHTS: &[u32] = &[
     4194304, 4, 128, 2097152, 8388608, 4194304, 131072, 524288, 8388608, 2097152, 8388608, 131072,
     2097152, 524288, 4194304, 32768, 262144,
 ];
